Refactors slack plugins views.py
