+++ src/dispatch/plugins/dispatch_slack/views.py
@@ -1,5 +1,3 @@
-import arrow
-import datetime
 import hashlib
 import hmac
 import json
@@ -8,71 +6,29 @@ import platform
 import sys
 
 from time import time
-from typing import List
 
 from fastapi import APIRouter, BackgroundTasks, Depends, Header, HTTPException
-
-from pydantic import BaseModel
 
 from sqlalchemy.orm import Session
 
 from starlette.requests import Request
 from starlette.responses import Response
 
-from dispatch.config import INCIDENT_PLUGIN_CONTACT_SLUG, INCIDENT_PLUGIN_CONVERSATION_SLUG
-from dispatch.conversation.enums import ConversationButtonActions
+from dispatch.database import get_db
 from dispatch.conversation.service import get_by_channel_id
-from dispatch.database import get_db, SessionLocal
-from dispatch.decorators import background_task
-from dispatch.enums import Visibility
-from dispatch.event import service as event_service
-from dispatch.incident import flows as incident_flows
-from dispatch.incident import service as incident_service
-from dispatch.incident.enums import IncidentStatus
-from dispatch.incident.models import IncidentUpdate, IncidentRead
-from dispatch.incident_priority import service as incident_priority_service
-from dispatch.incident_type import service as incident_type_service
-from dispatch.individual import service as individual_service
-from dispatch.participant import service as participant_service
-from dispatch.participant_role import service as participant_role_service
-from dispatch.participant_role.models import ParticipantRoleType
-from dispatch.plugins.base import plugins
 from dispatch.plugins.dispatch_slack import service as dispatch_slack_service
-from dispatch.report import flows as report_flows
-from dispatch.report import service as report_service
-from dispatch.report.enums import ReportTypes
-from dispatch.service import service as service_service
-from dispatch.task import service as task_service
-from dispatch.task.models import TaskStatus, Task
 
 from . import __version__
-from .config import (
-    SLACK_COMMAND_ASSIGN_ROLE_SLUG,
-    SLACK_COMMAND_ENGAGE_ONCALL_SLUG,
-    SLACK_COMMAND_EXECUTIVE_REPORT_SLUG,
-    SLACK_COMMAND_LIST_PARTICIPANTS_SLUG,
-    SLACK_COMMAND_LIST_RESOURCES_SLUG,
-    SLACK_COMMAND_LIST_TASKS_SLUG,
-    SLACK_COMMAND_LIST_MY_TASKS_SLUG,
-    SLACK_COMMAND_REPORT_INCIDENT_SLUG,
-    SLACK_COMMAND_TACTICAL_REPORT_SLUG,
-    SLACK_COMMAND_UPDATE_INCIDENT_SLUG,
-    SLACK_COMMAND_UPDATE_PARTICIPANT_SLUG,
-    SLACK_SIGNING_SECRET,
-    SLACK_TIMELINE_EVENT_REACTION,
-    SLACK_BAN_THREADS,
-)
+from .config import SLACK_SIGNING_SECRET
 from .messaging import (
     INCIDENT_CONVERSATION_COMMAND_MESSAGE,
     render_non_incident_conversation_command_error_message,
 )
 
-from .service import get_user_email
-from .modals import (
-    handle_modal_action,
-    create_update_participant_modal,
-    create_report_incident_modal,
-)
+from .modals import handle_modal_action
+from .commands import command_functions
+from .events import event_functions, get_channel_id_from_event, EventEnvelope
+from .actions import handle_block_action, handle_dialog_action
 
 
 router = APIRouter()
@@ -82,716 +38,6 @@ log = logging.getLogger(__name__)
 
 class SlackEventAppException(Exception):
     pass
-
-
-class EventBodyItem(BaseModel):
-    """Body item of the Slack event."""
-
-    type: str = None
-    channel: str = None
-    ts: str = None
-
-
-class EventBody(BaseModel):
-    """Body of the Slack event."""
-
-    channel: str = None
-    channel_id: str = None
-    channel_type: str = None
-    deleted_ts: str = None
-    event_ts: str = None
-    thread_ts: str = None
-    file_id: str = None
-    hidden: bool = None
-    inviter: str = None
-    item: EventBodyItem = None
-    item_user: str = None
-    reaction: str = None
-    subtype: str = None
-    team: str = None
-    text: str = None
-    type: str
-    user: str = None
-    user_id: str = None
-
-
-class EventEnvelope(BaseModel):
-    """Envelope of the Slack event."""
-
-    api_app_id: str = None
-    authed_users: List[str] = []
-    challenge: str = None
-    enterprise_id: str = None
-    event: EventBody = None
-    event_id: str = None
-    event_time: int = None
-    team_id: str = None
-    token: str = None
-    type: str
-
-
-@background_task
-def handle_reaction_added_event(
-    user_email: str, incident_id: int, event: dict = None, db_session=None
-):
-    """Handles an event where a reaction is added to a message."""
-    reaction = event.event.reaction
-
-    if reaction == SLACK_TIMELINE_EVENT_REACTION:
-        conversation_id = event.event.item.channel
-        message_ts = event.event.item.ts
-        message_ts_utc = datetime.datetime.utcfromtimestamp(float(message_ts))
-
-        # we fetch the message information
-        response = dispatch_slack_service.list_conversation_messages(
-            slack_client, conversation_id, latest=message_ts, limit=1, inclusive=1
-        )
-        message_text = response["messages"][0]["text"]
-        message_sender_id = response["messages"][0]["user"]
-
-        # we fetch the individual who sent the message
-        message_sender_email = get_user_email(client=slack_client, user_id=message_sender_id)
-        individual = individual_service.get_by_email(
-            db_session=db_session, email=message_sender_email
-        )
-
-        convo_plugin = plugins.get(INCIDENT_PLUGIN_CONVERSATION_SLUG)
-
-        # we log the event
-        event_service.log(
-            db_session=db_session,
-            source=convo_plugin.title,
-            description=f'"{message_text}," said {individual.name}',
-            incident_id=incident_id,
-            individual_id=individual.id,
-            started_at=message_ts_utc,
-        )
-
-
-@background_task
-def handle_reaction_removed_event(
-    user_email: str, incident_id: int, event: dict = None, db_session=None
-):
-    """Handles an event where a reaction is removed from a message."""
-    pass
-
-
-@background_task
-def add_evidence_to_storage(user_email: str, incident_id: int, event: dict = None, db_session=None):
-    """Adds evidence (e.g. files) added/shared in the conversation to storage."""
-    pass
-
-
-def is_business_hours(commander_tz: str):
-    """Determines if it's currently office hours where the incident commander is located."""
-    now = arrow.utcnow().to(commander_tz)
-    return now.weekday() not in [5, 6] and 9 <= now.hour < 17
-
-
-@background_task
-def after_hours(user_email: str, incident_id: int, event: dict = None, db_session=None):
-    """Notifies the user that this incident is current in after hours mode."""
-    incident = incident_service.get(db_session=db_session, incident_id=incident_id)
-
-    # get their timezone from slack
-    commander_info = dispatch_slack_service.get_user_info_by_email(
-        slack_client, email=incident.commander.email
-    )
-
-    commander_tz = commander_info["tz"]
-
-    if not is_business_hours(commander_tz):
-        # send ephermal message
-        blocks = [
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": (
-                        (
-                            f"Responses may be delayed. The current incident priority is *{incident.incident_priority.name}*"
-                            f" and your message was sent outside of the Incident Commander's working hours (Weekdays, 9am-5pm, {commander_tz} timezone)."
-                        )
-                    ),
-                },
-            }
-        ]
-
-        participant = participant_service.get_by_incident_id_and_email(
-            incident_id=incident_id, email=user_email
-        )
-        if not participant.after_hours_notification:
-            user_id = dispatch_slack_service.resolve_user(slack_client, user_email)["id"]
-            dispatch_slack_service.send_ephemeral_message(
-                slack_client, incident.conversation.channel_id, user_id, "", blocks=blocks
-            )
-            participant.after_hours_notification = True
-            db_session.add(participant)
-            db_session.commit()
-
-
-def filter_tasks_by_assignee_and_creator(tasks: List[Task], by_assignee: str, by_creator: str):
-    """Filters a list of tasks looking for a given creator or assignee."""
-    filtered_tasks = []
-    for t in tasks:
-        if by_creator:
-            creator_email = t.creator.individual.email
-            if creator_email == by_creator:
-                filtered_tasks.append(t)
-
-        if by_assignee:
-            assignee_emails = [a.individual.email for a in t.assignees]
-            if by_assignee in assignee_emails:
-                filtered_tasks.append(t)
-
-    return filtered_tasks
-
-
-@background_task
-def list_my_tasks(incident_id: int, command: dict = None, db_session=None):
-    """Returns the list of incident tasks to the user as an ephemeral message."""
-    user_email = dispatch_slack_service.get_user_email(slack_client, command["user_id"])
-    list_tasks(
-        incident_id=incident_id,
-        command=command,
-        db_session=db_session,
-        by_creator=user_email,
-        by_assignee=user_email,
-    )
-
-
-@background_task
-def list_tasks(
-    incident_id: int,
-    command: dict = None,
-    db_session=None,
-    by_creator: str = None,
-    by_assignee: str = None,
-):
-    """Returns the list of incident tasks to the user as an ephemeral message."""
-    blocks = []
-    for status in TaskStatus:
-        blocks.append(
-            {
-                "type": "section",
-                "text": {"type": "mrkdwn", "text": f"*{status.value} Incident Tasks*"},
-            }
-        )
-
-        tasks = task_service.get_all_by_incident_id_and_status(
-            db_session=db_session, incident_id=incident_id, status=status.value
-        )
-
-        if by_creator or by_assignee:
-            tasks = filter_tasks_by_assignee_and_creator(tasks, by_assignee, by_creator)
-
-        for task in tasks:
-            assignees = [a.individual.email for a in task.assignees]
-            blocks.append(
-                {
-                    "type": "section",
-                    "text": {
-                        "type": "mrkdwn",
-                        "text": (
-                            f"*Description:* <{task.weblink}|{task.description}>\n"
-                            f"*Assignees:* {', '.join(assignees)}"
-                        ),
-                    },
-                }
-            )
-        blocks.append({"type": "divider"})
-
-    dispatch_slack_service.send_ephemeral_message(
-        slack_client,
-        command["channel_id"],
-        command["user_id"],
-        "Incident List Tasks",
-        blocks=blocks,
-    )
-
-
-@background_task
-def list_participants(incident_id: int, command: dict = None, db_session=None):
-    """Returns the list of incident participants to the user as an ephemeral message."""
-    blocks = []
-    blocks.append(
-        {"type": "section", "text": {"type": "mrkdwn", "text": "*Incident Participants*"}}
-    )
-
-    participants = participant_service.get_all_by_incident_id(
-        db_session=db_session, incident_id=incident_id
-    ).all()
-
-    contact_plugin = plugins.get(INCIDENT_PLUGIN_CONTACT_SLUG)
-
-    for participant in participants:
-        if participant.is_active:
-            participant_email = participant.individual.email
-            participant_info = contact_plugin.get(participant_email)
-            participant_name = participant_info["fullname"]
-            participant_team = participant_info["team"]
-            participant_department = participant_info["department"]
-            participant_location = participant_info["location"]
-            participant_weblink = participant_info["weblink"]
-            participant_avatar_url = dispatch_slack_service.get_user_avatar_url(
-                slack_client, participant_email
-            )
-
-            participant_reason_added = participant.added_reason or "Unknown"
-            if participant.added_by:
-                participant_added_by = participant.added_by.individual.name
-            else:
-                participant_added_by = "Unknown"
-
-            participant_active_roles = participant_role_service.get_all_active_roles(
-                db_session=db_session, participant_id=participant.id
-            )
-            participant_roles = []
-            for role in participant_active_roles:
-                participant_roles.append(role.role)
-
-            block = {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": (
-                        f"*Name:* <{participant_weblink}|{participant_name}>\n"
-                        f"*Team*: {participant_team}, {participant_department}\n"
-                        f"*Location*: {participant_location}\n"
-                        f"*Incident Role(s)*: {(', ').join(participant_roles)}\n"
-                        f"*Reason Added*: {participant_reason_added}\n"
-                        f"*Added By*: {participant_added_by}\n"
-                    ),
-                },
-            }
-
-            if len(participants) < 20:
-                block.update(
-                    {
-                        "accessory": {
-                            "type": "image",
-                            "alt_text": participant_name,
-                            "image_url": participant_avatar_url,
-                        },
-                    }
-                )
-
-            blocks.append(block)
-            blocks.append({"type": "divider"})
-
-    dispatch_slack_service.send_ephemeral_message(
-        slack_client,
-        command["channel_id"],
-        command["user_id"],
-        "Incident List Participants",
-        blocks=blocks,
-    )
-
-
-def create_assign_role_dialog(incident_id: int, command: dict = None):
-    """Creates a dialog for assigning a role."""
-    role_options = []
-    for role in ParticipantRoleType:
-        if role != ParticipantRoleType.participant:
-            role_options.append({"label": role.value, "value": role.value})
-
-    dialog = {
-        "callback_id": command["command"],
-        "title": "Assign Role",
-        "submit_label": "Assign",
-        "elements": [
-            {
-                "label": "Participant",
-                "type": "select",
-                "name": "participant",
-                "data_source": "users",
-            },
-            {"label": "Role", "type": "select", "name": "role", "options": role_options},
-        ],
-    }
-
-    dispatch_slack_service.open_dialog_with_user(slack_client, command["trigger_id"], dialog)
-
-
-@background_task
-def create_update_incident_dialog(incident_id: int, command: dict = None, db_session=None):
-    """Creates a dialog for updating incident information."""
-    incident = incident_service.get(db_session=db_session, incident_id=incident_id)
-
-    type_options = []
-    for t in incident_type_service.get_all(db_session=db_session):
-        type_options.append({"label": t.name, "value": t.name})
-
-    priority_options = []
-    for priority in incident_priority_service.get_all(db_session=db_session):
-        priority_options.append({"label": priority.name, "value": priority.name})
-
-    status_options = []
-    for status in IncidentStatus:
-        status_options.append({"label": status.value, "value": status.value})
-
-    visibility_options = []
-    for visibility in Visibility:
-        visibility_options.append({"label": visibility.value, "value": visibility.value})
-
-    notify_options = [{"label": "Yes", "value": "Yes"}, {"label": "No", "value": "No"}]
-
-    dialog = {
-        "callback_id": command["command"],
-        "title": "Update Incident",
-        "submit_label": "Save",
-        "elements": [
-            {"type": "textarea", "label": "Title", "name": "title", "value": incident.title},
-            {
-                "type": "textarea",
-                "label": "Description",
-                "name": "description",
-                "value": incident.description,
-            },
-            {
-                "label": "Type",
-                "type": "select",
-                "name": "type",
-                "value": incident.incident_type.name,
-                "options": type_options,
-            },
-            {
-                "label": "Priority",
-                "type": "select",
-                "name": "priority",
-                "value": incident.incident_priority.name,
-                "options": priority_options,
-            },
-            {
-                "label": "Status",
-                "type": "select",
-                "name": "status",
-                "value": incident.status,
-                "options": status_options,
-            },
-            {
-                "label": "Visibility",
-                "type": "select",
-                "name": "visibility",
-                "value": incident.visibility,
-                "options": visibility_options,
-            },
-            {
-                "label": "Notify on change",
-                "type": "select",
-                "name": "notify",
-                "value": "Yes",
-                "options": notify_options,
-            },
-        ],
-    }
-
-    dispatch_slack_service.open_dialog_with_user(slack_client, command["trigger_id"], dialog)
-
-
-@background_task
-def create_engage_oncall_dialog(incident_id: int, command: dict = None, db_session=None):
-    """Creates a dialog to engage an oncall person."""
-    oncall_services = service_service.get_all_by_status(db_session=db_session, is_active=True)
-
-    if not oncall_services.count():
-        blocks = [
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "No oncall services have been defined. You can define them in the Dispatch UI at /services",
-                },
-            }
-        ]
-        dispatch_slack_service.send_ephemeral_message(
-            slack_client,
-            command["channel_id"],
-            command["user_id"],
-            "No Oncall Services Defined",
-            blocks=blocks,
-        )
-        return
-
-    oncall_service_options = []
-    for oncall_service in oncall_services:
-        oncall_service_options.append(
-            {"label": oncall_service.name, "value": oncall_service.external_id}
-        )
-
-    page_options = [{"label": "Yes", "value": "Yes"}, {"label": "No", "value": "No"}]
-
-    dialog = {
-        "callback_id": command["command"],
-        "title": "Engage Oncall",
-        "submit_label": "Engage",
-        "elements": [
-            {
-                "label": "Oncall Service",
-                "type": "select",
-                "name": "oncall_service_id",
-                "options": oncall_service_options,
-            },
-            {
-                "label": "Page",
-                "type": "select",
-                "name": "page",
-                "value": "No",
-                "options": page_options,
-            },
-        ],
-    }
-
-    dispatch_slack_service.open_dialog_with_user(slack_client, command["trigger_id"], dialog)
-
-
-@background_task
-def create_tactical_report_dialog(incident_id: int, command: dict = None, db_session=None):
-    """Creates a dialog with the most recent tactical report data, if it exists."""
-    # we load the most recent tactical report
-    tactical_report = report_service.get_most_recent_by_incident_id_and_type(
-        db_session=db_session, incident_id=incident_id, report_type=ReportTypes.tactical_report
-    )
-
-    conditions = actions = needs = ""
-    if tactical_report:
-        conditions = tactical_report.details.get("conditions")
-        actions = tactical_report.details.get("actions")
-        needs = tactical_report.details.get("needs")
-
-    dialog = {
-        "callback_id": command["command"],
-        "title": "Tactical Report",
-        "submit_label": "Submit",
-        "elements": [
-            {"type": "textarea", "label": "Conditions", "name": "conditions", "value": conditions},
-            {"type": "textarea", "label": "Actions", "name": "actions", "value": actions},
-            {"type": "textarea", "label": "Needs", "name": "needs", "value": needs},
-        ],
-    }
-
-    dispatch_slack_service.open_dialog_with_user(slack_client, command["trigger_id"], dialog)
-
-
-@background_task
-def create_executive_report_dialog(incident_id: int, command: dict = None, db_session=None):
-    """Creates a dialog with the most recent executive report data, if it exists."""
-    # we load the most recent executive report
-    executive_report = report_service.get_most_recent_by_incident_id_and_type(
-        db_session=db_session, incident_id=incident_id, report_type=ReportTypes.executive_report
-    )
-
-    current_status = overview = next_steps = ""
-    if executive_report:
-        current_status = executive_report.details.get("current_status")
-        overview = executive_report.details.get("overview")
-        next_steps = executive_report.details.get("next_steps")
-
-    dialog = {
-        "callback_id": command["command"],
-        "title": "Executive Report",
-        "submit_label": "Submit",
-        "elements": [
-            {
-                "type": "textarea",
-                "label": "Current Status",
-                "name": "current_status",
-                "value": current_status,
-            },
-            {"type": "textarea", "label": "Overview", "name": "overview", "value": overview},
-            {"type": "textarea", "label": "Next Steps", "name": "next_steps", "value": next_steps},
-        ],
-    }
-
-    dispatch_slack_service.open_dialog_with_user(slack_client, command["trigger_id"], dialog)
-
-
-@background_task
-def add_user_to_conversation(
-    user_id: str, user_email: str, incident_id: int, action: dict, db_session=None
-):
-    """Adds a user to a conversation."""
-    incident = incident_service.get(db_session=db_session, incident_id=incident_id)
-
-    if incident.status == IncidentStatus.closed:
-        message = f"Sorry, we cannot add you to a closed incident. Please reach out to the incident commander ({incident.commander.name}) for details."
-        dispatch_slack_service.send_ephemeral_message(
-            slack_client, action["container"]["channel_id"], user_id, message
-        )
-    else:
-        dispatch_slack_service.add_users_to_conversation(
-            slack_client, incident.conversation.channel_id, [user_id]
-        )
-        message = f"Success! We've added you to incident {incident.name}. Please check your side bar for the new channel."
-        dispatch_slack_service.send_ephemeral_message(
-            slack_client, action["container"]["channel_id"], user_id, message
-        )
-
-
-@background_task
-def member_joined_channel(
-    user_email: str, incident_id: int, event: EventEnvelope, db_session=None,
-):
-    """Handles the member_joined_channel slack event."""
-    participant = incident_flows.incident_add_or_reactivate_participant_flow(
-        user_email=user_email, incident_id=incident_id, db_session=db_session
-    )
-
-    # update participant metadata
-    inviter_email = get_user_email(client=slack_client, user_id=event.event.inviter)
-    added_by_participant = participant_service.get_by_incident_id_and_email(
-        db_session=db_session, incident_id=incident_id, email=inviter_email
-    )
-    participant.added_by = added_by_participant
-    participant.added_reason = event.event.text
-
-    db_session.commit()
-
-
-@background_task
-def ban_threads_warning(user_email: str, incident_id: int, event: dict = None, db_session=None):
-    """Sends the user an ephemeral message if they use threads."""
-    if not SLACK_BAN_THREADS:
-        return
-
-    if event.event.thread_ts:
-        # we should be able to look for `subtype == message_replied` once this bug is fixed
-        # https://api.slack.com/events/message/message_replied
-        # From Slack: Bug alert! This event is missing the subtype field when dispatched
-        # over the Events API. Until it is fixed, examine message events' thread_ts value.
-        # When present, it's a reply. To be doubly sure, compare a thread_ts to the top-level ts
-        # value, when they differ the latter is a reply to the former.
-        message = "Please refrain from using threads in incident related channels. Threads make it harder for incident participants to maintain context."
-        dispatch_slack_service.send_ephemeral_message(
-            slack_client,
-            event.event.channel,
-            event.event.user,
-            message,
-            thread_ts=event.event.thread_ts,
-        )
-
-
-def event_functions(event: EventEnvelope):
-    """Interprets the events and routes it the appropriate function."""
-    event_mappings = {
-        "file_created": [add_evidence_to_storage],
-        "file_shared": [add_evidence_to_storage],
-        "link_shared": [],
-        "member_joined_channel": [member_joined_channel],
-        "message": [ban_threads_warning],
-        "member_left_channel": [incident_flows.incident_remove_participant_flow],
-        "message.groups": [],
-        "message.im": [],
-        "reaction_added": [handle_reaction_added_event],
-        "reaction_removed": [handle_reaction_removed_event],
-    }
-
-    return event_mappings.get(event.event.type, [])
-
-
-def command_functions(command: str):
-    """Interprets the command and routes it the appropriate function."""
-    command_mappings = {
-        SLACK_COMMAND_ASSIGN_ROLE_SLUG: [create_assign_role_dialog],
-        SLACK_COMMAND_ENGAGE_ONCALL_SLUG: [create_engage_oncall_dialog],
-        SLACK_COMMAND_EXECUTIVE_REPORT_SLUG: [create_executive_report_dialog],
-        SLACK_COMMAND_LIST_PARTICIPANTS_SLUG: [list_participants],
-        SLACK_COMMAND_LIST_RESOURCES_SLUG: [incident_flows.incident_list_resources_flow],
-        SLACK_COMMAND_LIST_TASKS_SLUG: [list_tasks],
-        SLACK_COMMAND_LIST_MY_TASKS_SLUG: [list_my_tasks],
-        SLACK_COMMAND_TACTICAL_REPORT_SLUG: [create_tactical_report_dialog],
-        SLACK_COMMAND_UPDATE_INCIDENT_SLUG: [create_update_incident_dialog],
-        SLACK_COMMAND_UPDATE_PARTICIPANT_SLUG: [create_update_participant_modal],
-        SLACK_COMMAND_REPORT_INCIDENT_SLUG: [create_report_incident_modal],
-    }
-
-    return command_mappings.get(command, [])
-
-
-@background_task
-def handle_update_incident_action(user_id, user_email, incident_id, action, db_session=None):
-    """Messages slack dialog data into something that Dispatch can use."""
-    submission = action["submission"]
-    notify = True if submission["notify"] == "Yes" else False
-    incident_in = IncidentUpdate(
-        title=submission["title"],
-        description=submission["description"],
-        incident_type={"name": submission["type"]},
-        incident_priority={"name": submission["priority"]},
-        status=submission["status"],
-        visibility=submission["visibility"],
-    )
-
-    incident = incident_service.get(db_session=db_session, incident_id=incident_id)
-    existing_incident = IncidentRead.from_orm(incident)
-    incident_service.update(db_session=db_session, incident=incident, incident_in=incident_in)
-    incident_flows.incident_update_flow(user_email, incident_id, existing_incident, notify)
-
-
-@background_task
-def handle_assign_role_action(user_id, user_email, incident_id, action, db_session=None):
-    """Messages slack dialog data into some thing that Dispatch can use."""
-    assignee_user_id = action["submission"]["participant"]
-    assignee_role = action["submission"]["role"]
-    assignee_email = get_user_email(client=slack_client, user_id=assignee_user_id)
-    incident_flows.incident_assign_role_flow(user_email, incident_id, assignee_email, assignee_role)
-
-
-def dialog_action_functions(action: str):
-    """Interprets the action and routes it to the appropriate function."""
-    action_mappings = {
-        SLACK_COMMAND_ASSIGN_ROLE_SLUG: [handle_assign_role_action],
-        SLACK_COMMAND_ENGAGE_ONCALL_SLUG: [incident_flows.incident_engage_oncall_flow],
-        SLACK_COMMAND_EXECUTIVE_REPORT_SLUG: [report_flows.create_executive_report],
-        SLACK_COMMAND_TACTICAL_REPORT_SLUG: [report_flows.create_tactical_report],
-        SLACK_COMMAND_UPDATE_INCIDENT_SLUG: [handle_update_incident_action],
-    }
-
-    # this allows for unique action blocks e.g. invite-user or invite-user-1, etc
-    for key in action_mappings.keys():
-        if key in action:
-            return action_mappings[key]
-    return []
-
-
-def block_action_functions(action: str):
-    """Interprets the action and routes it to the appropriate function."""
-    action_mappings = {
-        ConversationButtonActions.invite_user: [add_user_to_conversation],
-    }
-
-    # this allows for unique action blocks e.g. invite-user or invite-user-1, etc
-    for key in action_mappings.keys():
-        if key in action:
-            return action_mappings[key]
-    return []
-
-
-def handle_dialog_action(action: dict, background_tasks: BackgroundTasks, db_session: SessionLocal):
-    """Handles all dialog actions."""
-    channel_id = action["channel"]["id"]
-    conversation = get_by_channel_id(db_session=db_session, channel_id=channel_id)
-    incident_id = conversation.incident_id
-
-    user_id = action["user"]["id"]
-    user_email = action["user"]["email"]
-
-    action_id = action["callback_id"]
-
-    for f in dialog_action_functions(action_id):
-        background_tasks.add_task(f, user_id, user_email, incident_id, action)
-
-
-def handle_block_action(action: dict, background_tasks: BackgroundTasks):
-    """Handles a standalone block action."""
-
-    action_id = action["actions"][0]["block_id"]
-    incident_id = action["actions"][0]["value"]
-    user_id = action["user"]["id"]
-    user_email = action["user"]["email"]
-
-    for f in block_action_functions(action_id):
-        background_tasks.add_task(f, user_id, user_email, incident_id, action)
 
 
 def create_ua_string():
@@ -828,19 +74,6 @@ def verify_timestamp(timestamp: int):
         raise HTTPException(status_code=403, detail="Invalid request timestamp")
 
 
-def get_channel_id(event_body: dict):
-    """Returns the channel id from the Slack event."""
-    channel_id = ""
-    if event_body.channel_id:
-        return event_body.channel_id
-    if event_body.channel:
-        return event_body.channel
-    if event_body.item.channel:
-        return event_body.item.channel
-
-    return channel_id
-
-
 @router.post("/slack/event")
 async def handle_event(
     event: EventEnvelope,
@@ -867,7 +100,7 @@ async def handle_event(
     event_body = event.event
 
     user_id = event_body.user
-    channel_id = get_channel_id(event_body)
+    channel_id = get_channel_id_from_event(event_body)
 
     if user_id and channel_id:
         conversation = get_by_channel_id(db_session=db_session, channel_id=channel_id)
